+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v0.3.0
+### Added
+* superlineages can optionally be based on data from cov-lineages
+* date info can be added in for plotting lineage appearance
+
 ## v0.2.0
 ### Changed
 * lineage plot now has lineages labeled